server/requirements.txt
defusedxml==0.4.1
Django==1.9.5
django-cors-headers==1.1.0
django-filter==0.13.0
django-smtp-ssl==1.0
django-websocket-redis==0.4.6
django-wysiwyg-redactor==0.4.9
djangorestframework==3.3.3
gevent==1.1.1
greenlet==0.4.9
oauthlib==1.0.3
Pillow==3.2.0
psycopg2==2.6.1
PyJWT==1.4.0
pypng==0.0.18
python-social-auth==0.2.14
python3-openid==3.0.10
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
sorl-thumbnail==12.3
uWSGI==2.0.12
